[Core] Move ray_utils.py from `engine` to `executor` package (#4347)
